Fixed Bug: Listeners were not deactivated and added multiple times
processDeactivateion referenced the wrong map, therefore, listeners were
not removed from the set of active listeners

On top of that, the set of active listeners used to be a list of active
listeners, since listeners were not deactivated properly, they were
added multiple times to the list.

This resulted in e.g. trips being aggregated multiple times.

The wrongly referenced map was replaced and the list of active listeners
was replaced by a set.